tests/wpt/tests/html/canvas/element/pixel-manipulation/2d.imageData.object.ctor.pixelFormat.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.imageData.object.ctor.pixelFormat</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.imageData.object.ctor.pixelFormat</h1>
<p class="desc">Testing ImageData constructor pixelFormat behaviors</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Testing ImageData constructor pixelFormat behaviors");
_addTest(function(canvas, ctx) {

  function setRGBA(imageData, i, rgba)
  {
      var s = i * 4;
      imageData[s] = rgba[0];
      imageData[s + 1] = rgba[1];
      imageData[s + 2] = rgba[2];
      imageData[s + 3] = rgba[3];
  }

  function getRGBA(imageData, i)
  {
      var result = [];
      var s = i * 4;
      for (var j = 0; j < 4; j++) {
          result[j] = imageData[s + j];
      }
      return result;
  }

  function assertArrayEquals(actual, expected)
  {
      _assertSame(typeof actual, "object", "typeof actual", "\"object\"");
      _assertDifferent(actual, null, "actual", "null");
      _assertSame("length" in actual, true, "\"length\" in actual", "true");
      _assertSame(actual.length, expected.length, "actual.length", "expected.length");
      for (var i = 0; i < actual.length; i++) {
          _assertSame(actual.hasOwnProperty(i), expected.hasOwnProperty(i), "actual.hasOwnProperty(i)", "expected.hasOwnProperty(i)");
          _assertSame(actual[i], expected[i], "actual[\""+(i)+"\"]", "expected[\""+(i)+"\"]");
      }
  }

  _assertDifferent(ImageData, undefined, "ImageData", "undefined");
  imageData = new ImageData(100, 50);
  _assertSame(imageData.pixelFormat, "rgba-unorm8", "imageData.pixelFormat", "\"rgba-unorm8\"");

  imageData = new ImageData(100, 50, {pixelFormat:"rgba-float16"});
  _assertSame(imageData.pixelFormat, "rgba-float16", "imageData.pixelFormat", "\"rgba-float16\"");
  var testColor = [0, -1, 0.5, 1024];
  setRGBA(imageData.data, 4, testColor);
  assertArrayEquals(getRGBA(imageData.data, 4), testColor);

  data = new Float16Array(200);
  imageData = new ImageData(data, 10, 5, {pixelFormat:"rgba-float16"});
  _assertSame(imageData.pixelFormat, "rgba-float16", "imageData.pixelFormat", "\"rgba-float16\"");
  _assertSame(imageData.data, data, "imageData.data", "data");
  assert_throws_dom("INVALID_STATE_ERR", function() { new ImageData(data, 10, 5, {pixelFormat:"rgba-unorm8"}); });
  assert_throws_dom("INVALID_STATE_ERR", function() { new ImageData(data, 10, 5); });

  data = new Uint8ClampedArray(200);
  imageData = new ImageData(data, 10, 5);
  _assertSame(imageData.pixelFormat, "rgba-unorm8", "imageData.pixelFormat", "\"rgba-unorm8\"");
  _assertSame(imageData.data, data, "imageData.data", "data");
  assert_throws_dom("INVALID_STATE_ERR", function() { new ImageData(data, 10, 5, {pixelFormat:"rgba-float16"}); });

  assert_throws_js(TypeError, function() { new ImageData(data, 10, 5, {pixelFormat:"unorm8"}); });
  assert_throws_js(TypeError, function() { new ImageData(data, 10, 5, {pixelFormat:"float16"}); });
  assert_throws_js(TypeError, function() { new ImageData(data, 10, 5, {pixelFormat:"rgba8unorm"}); });
  assert_throws_js(TypeError, function() { new ImageData(data, 10, 5, {pixelFormat:"rgba16float"}); });

});
</script>
